+++ demo/src/main/java/cn/bingoogolapple/refreshlayout/demo/ui/activity/NormalRecyclerViewActivity.java
@@ -61,7 +61,10 @@ public class NormalRecyclerViewActivity extends BaseActivity implements BGAOnRVI
 
     @Override
     protected void processLogic(Bundle savedInstanceState) {
-        mRefreshLayout.setRefreshViewHolder(new BGAMoocStyleRefreshViewHolder(mApp, true));
+        BGAMoocStyleRefreshViewHolder moocStyleRefreshViewHolder = new BGAMoocStyleRefreshViewHolder(mApp, true);
+        moocStyleRefreshViewHolder.setOriginalImage(R.mipmap.bga_refresh_moooc);
+        moocStyleRefreshViewHolder.setUltimateColor(R.color.imoocstyle);
+        mRefreshLayout.setRefreshViewHolder(moocStyleRefreshViewHolder);
 
         initBanner();